Extracted CodeGen concerns to a separate entrypoint
