+++ CustomCADs.Staging.sln
@@ -115,10 +115,12 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Shared.Speedy", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Shared.UseCases", "src\Shared\UseCases\CustomCADs.Shared.UseCases.csproj", "{00401D1F-828D-4A1D-B363-A06B8604EBFA}"
 EndProject
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{F67E7870-8219-9EC9-916E-B9F4BD6F71AB}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Migrations", "src\Tools\Migrations\CustomCADs.Migrations.csproj", "{2AAB9F36-14E5-4786-9B30-D11B14117FD5}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Tools.CodeGen", "src\Tools\CodeGen\CustomCADs.Tools.CodeGen.csproj", "{B493FBDD-10AA-430F-84C8-A15AD0E762F2}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -669,6 +671,18 @@ Global
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5}.Release|x64.Build.0 = Release|Any CPU
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5}.Release|x86.ActiveCfg = Release|Any CPU
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5}.Release|x86.Build.0 = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x64.Build.0 = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x86.Build.0 = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x64.ActiveCfg = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x64.Build.0 = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x86.ActiveCfg = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x86.Build.0 = Release|Any CPU	
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -731,5 +745,6 @@ Global
 		{00401D1F-828D-4A1D-B363-A06B8604EBFA} = {C1CBD8F6-DA83-433C-91B3-4AE8C9DF8A10}
 		{F67E7870-8219-9EC9-916E-B9F4BD6F71AB} = {74636A63-AB16-4950-BF0D-966FBFF79B83}
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5} = {F67E7870-8219-9EC9-916E-B9F4BD6F71AB}
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2} = {F67E7870-8219-9EC9-916E-B9F4BD6F71AB}
 	EndGlobalSection
 EndGlobal
