+++ CustomCADs.sln
@@ -119,6 +119,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tools", "Tools", "{F67E7870
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Tools.Migrations", "src\Tools\Migrations\CustomCADs.Tools.Migrations.csproj", "{2AAB9F36-14E5-4786-9B30-D11B14117FD5}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CustomCADs.Tools.CodeGen", "src\Tools\CodeGen\CustomCADs.Tools.CodeGen.csproj", "{B493FBDD-10AA-430F-84C8-A15AD0E762F2}"
+EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{6A7D85BE-786A-4537-8454-3FC6C88C57A0}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "UnitTests", "UnitTests", "{E27610CC-5AA4-4B13-9DCF-E77B340A4749}"
@@ -796,6 +798,18 @@ Global
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5}.Release|x64.Build.0 = Release|Any CPU
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5}.Release|x86.ActiveCfg = Release|Any CPU
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5}.Release|x86.Build.0 = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x64.Build.0 = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Debug|x86.Build.0 = Debug|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x64.ActiveCfg = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x64.Build.0 = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x86.ActiveCfg = Release|Any CPU
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2}.Release|x86.Build.0 = Release|Any CPU	
 		{B47EF3A2-AE9F-4209-826F-D54353328EBC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B47EF3A2-AE9F-4209-826F-D54353328EBC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B47EF3A2-AE9F-4209-826F-D54353328EBC}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1146,6 +1160,7 @@ Global
 		{00401D1F-828D-4A1D-B363-A06B8604EBFA} = {C1CBD8F6-DA83-433C-91B3-4AE8C9DF8A10}
 		{F67E7870-8219-9EC9-916E-B9F4BD6F71AB} = {74636A63-AB16-4950-BF0D-966FBFF79B83}
 		{2AAB9F36-14E5-4786-9B30-D11B14117FD5} = {F67E7870-8219-9EC9-916E-B9F4BD6F71AB}
+		{B493FBDD-10AA-430F-84C8-A15AD0E762F2} = {F67E7870-8219-9EC9-916E-B9F4BD6F71AB}
 		{E27610CC-5AA4-4B13-9DCF-E77B340A4749} = {6A7D85BE-786A-4537-8454-3FC6C88C57A0}
 		{21E6ABAD-8ED7-4C5B-91AA-4713664006B7} = {E27610CC-5AA4-4B13-9DCF-E77B340A4749}
 		{B47EF3A2-AE9F-4209-826F-D54353328EBC} = {21E6ABAD-8ED7-4C5B-91AA-4713664006B7}
