+++ src/Tools/CodeGen/CodeGenReference.cs
@@ -1,0 +1,8 @@
+using System.Reflection;
+
+namespace CustomCADs.Tools.CodeGen;
+
+public static class CodeGenReference
+{
+	public static Assembly Assembly => typeof(CodeGenReference).Assembly;
+}
