+++ src/Tools/CodeGen/Program.cs
@@ -1,0 +1,22 @@
+using JasperFx;
+
+var builder = WebApplication.CreateBuilder(args);
+
+// Use Cases
+builder.Services.GenerateUseCases(builder.Environment);
+builder.Services.AddCacheService();
+builder.Services.AddBackgroundJobs();
+
+// External Services
+builder.Services.AddEmailService(builder.Configuration);
+builder.Services.AddTokensService(builder.Configuration);
+builder.Services.AddPaymentService(builder.Configuration);
+builder.Services.AddDeliveryService(builder.Configuration);
+builder.Services.AddStorageService(builder.Configuration);
+
+// Modules
+builder.Services.AddPersistence(builder.Configuration);
+builder.Services.AddIdentity();
+
+var app = builder.Build();
+return await app.RunJasperFxCommands(args).ConfigureAwait(false);
