+++ src/Tools/CodeGen/appsettings.json
@@ -1,0 +1,12 @@
+{
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "ApplicationConnection": "Host=localhost;Port=5432;Database=CustomCADs;Username=postgres;Password=change_this"
+  }
+}
